requirements.txt
discord.py[voice]
youtube_dl
PyNaCl